short: Allow existing links to be hijacked.
The edit form will be displayed instead of a plain error message if a user
attempts to create a short link owned by another user. The form will preserve
the existing input and adds another option to hijack the existing link.

This also adds a "Patches welcome" message to the landing page.