short: Add support for pattern-based redirects.
This is a quick experiment to look at the utility of being able to make, for
example, go/pr/12345 redirect to the relevant pull request on GitHub.

Comments can now be associated with a link to help clean up the front page.

The mapping logic is moved to an in-memory cache which is asynchronously
refreshed if the fingerprint of the underlying database table changes.