short: Add support for fallback URLs
Currently, if a pattern-substitution link doesn't match, there's no way to
provide any sort of fallback and the user is redirected to the default target.

This change allows a `||` separator to be used in a target URL to allow
multiple match attempts or to fall back to a default target.